Test folder * Move test package inside source code folder
