+++ src/test/cmp/infrastructure/controller/fixJSDOMPostMessage.js
