+++ src/test/integration/initializeTestClientCMP.js
