+++ src/test/resources/globalvendorlist.json
